Editar Perfil - Concluido

+++ app/Http/Controllers/TagsController.php
@@ -9,32 +9,18 @@ use Illuminate\Http\Request;
 
 class TagsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         return view('tags.index')->with('tags', Tag::all());
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         return view('tags.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(CreateTagsRequest $request)
     {
         Tag::create(['name' => $request->name]);
@@ -42,35 +28,16 @@ class TagsController extends Controller
         return redirect(route('tags.index'));
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Tag $tag)
     {
         return view('tags.edit')->with('tags', $tag);
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(EditTagsRequest $request, Tag $tag)
     {
         $tag->update([
@@ -80,16 +47,14 @@ class TagsController extends Controller
         return redirect(route('tags.index'));
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
         $tag = Tag::withTrashed()->where('id', $id)->firstOrFail();
-
+        $product = $tag->products()->count();
+        if ($product > 0) {
+            session()->flash('error', 'Esta tag possui (' . $product . ') produto(s) em uso e portanto não pode ser excluída!');
+            return redirect()->back();
+        }
         if ($tag->trashed()) {
             $tag->forceDelete();
             session()->flash('success', 'Tag apagada com sucesso!');
